group5-py-app/requirements.txt
# Framework Web et API
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Base de données et ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation de données
pydantic==2.5.0
pydantic-settings==2.1.0

# Requêtes HTTP
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Gestion des dates
python-dateutil==2.8.2

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
factory-boy==3.3.0

# Outils de développement
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx-autodoc-typehints==1.25.2

# Monitoring et logging
structlog==23.2.0
colorama==0.4.6

# Utilitaires
click==8.1.7
rich==13.7.0
typer==0.9.0

# Cache et performance
redis==5.0.1
cachetools==5.3.2

# Sérialisation
orjson==3.9.10

# Timezone
pytz==2023.3

# Authentification (optionnel)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Variables d'environnement et secrets
cryptography==41.0.7

# Outils mathématiques pour les recommandations
numpy==1.25.2
scipy==1.11.4

# Développement et debug
ipython==8.17.2
ipdb==0.13.13
sphinx
sphinx-autobuild
sphinxcontrib-napoleon